server/mailapi/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.97.0
requests==2.31.0
pydantic==2.3.0
python-multipart==0.0.6
itsdangerous==2.1.2
starlette==0.27.0 